backend/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.10
uuid==1.30
Werkzeug==1.0.1
